requirements.txt
pywinauto
pytest
pytest-html
logging
python-dotenv